Change sensor protocol

+++ Meteo/DataObjects/MeteoDto.cs
@@ -2,6 +2,6 @@
 
 public sealed record MeteoDto
 {
-    public double Temperature { get; init; }
+    public int Temperature { get; init; }
     public int Humidity { get; init; }
 }
